refactored w/ a new api layer
